p2p_test: topoMsg send to kafka

+++ p2p/config.go
@@ -36,7 +36,7 @@ const (
 	DefaultMaxPeers        uint      = 50
 	DefaultMaxPendingPeers uint      = 20
 	DefaultMaxInboundRatio uint      = 2
-	DefaultPort            uint16    = 8483
+	DefaultPort            uint      = 8483
 	DefaultNetID           NetworkID = Aquarius
 )
